feat: Add page-load module
- Add factory funtion page load to render main elements of ste
- Imported and exported pageLoad()
- Style module
- Also modify some scripts in package.json
- Also modify webpack.consig,js
- Also delete some unnecesary files

Refs. #1 and #2
