+++ package.json
@@ -5,7 +5,8 @@
   "private": true,
   "scripts": {
     "build": "webpack --config webpack.config.js --mode=development",
-    "watch": "nodemon dist/main.js",
+    "watch": "webpack --watch",
+    "start": "webpack-dev-server --open",
     "lint": "eslint .",
     "lint-fix": "eslint . --fix"
   },
@@ -34,10 +35,10 @@
     "eslint-config-airbnb": "^18.0.1",
     "eslint-plugin-import": "^2.20.1",
     "file-loader": "^5.1.0",
-    "nodemon": "^2.0.2",
     "style-loader": "^1.1.3",
+    "url-loader": "^3.0.0",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.11",
-    "webpack-node-externals": "^1.7.2"
+    "webpack-dev-server": "^3.10.3"
   }
 }
