+++ src/css/style.css
@@ -1,21 +1,134 @@
 body {
-  min-height: 100vh;
+  height: 100vh;
   scroll-behavior: smooth;
   text-rendering: optimizeSpeed;
   line-height: 1.5;
-  font-family: "Muli", sans-serif;
-  color: var(--text-color);
 }
 
 #content {
   display: grid;
   grid-template-columns: 1fr 1fr;
-  grid-template-rows: auto 50px;
+  grid-template-rows: auto auto 50px;
+  height: 100vh;
+}
+
+.main-content {
+  display: grid;
+  grid-template-rows: 5% auto;
+  padding: 2em;
+  background: linear-gradient(-45deg, #008c45, #f4f5f0, #cd212a);
+  background-size: 400% 400%;
+  animation: gradient 15s ease infinite;
+  text-align: center;
+}
+
+/* Left image */
+
+.restaurant-logo {
+  overflow: hidden;
+  background-size: cover;
+  background-position: center;
+  background-image: url("../img/lunch-table-5929.jpg");
   min-height: 100vh;
 }
 
+/* Navbar */
+nav {
+  grid-column-start: 1;
+  grid-column-end: 3;
+  grid-row-start: 1;
+  place-self: center;
+}
+
+ul {
+  list-style-type: none;
+}
+
+li {
+  display: inline-block;
+  cursor: pointer;
+  font-size: 1.25rem;
+  font-weight: 700;
+  margin: 0 1rem;
+  padding: 0 5px;
+}
+
+.links {
+  color: #0d5e23;
+  text-transform: uppercase;
+  text-decoration: none;
+  letter-spacing: 0.15em;
+  display: inline-block;
+  padding: 0px 20px;
+  position: relative;
+}
+
+.links:after {
+  background: none repeat scroll 0 0 transparent;
+  bottom: 0;
+  content: "";
+  display: block;
+  height: 2px;
+  left: 40%;
+  position: absolute;
+  background: #cd212a;
+  transition: width 0.2s ease 0s, left 0.2s ease 0s;
+  width: 0;
+}
+
+.links:hover:after {
+  width: 100%;
+  left: 0;
+}
+
+/*  Restaurant logo */
+
 img {
-  max-width: 100%;
-  max-height: 100%;
-  display: block; /* remove extra space below image */
+  position: relative;
+  margin: auto;
+  top: 8em;
+  width: 80%;
+}
+
+/* Footer */
+
+footer {
+  grid-column-start: 1;
+  grid-column-end: 3;
+  grid-row-start: 3;
+  place-self: center;
+  display: flex;
+}
+
+footer a {
+  text-decoration: none;
+}
+
+footer p {
+  color: #0d5e23;
+  font-size: 1rem;
+  font-weight: 700;
+}
+
+.copyright {
+  padding-left: 6px;
+}
+
+@keyframes gradient {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
+
+@media (max-width: 1000px) {
+  img {
+    top: 14em;
+    width: 100%;
+  }
 }
