+++ webpack.config.js
@@ -1,5 +1,4 @@
 const path = require('path');
-const nodeExternals = require('webpack-node-externals');
 
 module.exports = {
   target: 'node',
@@ -8,8 +7,9 @@ module.exports = {
     filename: 'main.js',
     path: path.resolve(__dirname, 'dist'),
   },
-  externals: [nodeExternals()],
-
+  devServer: {
+    contentBase: './dist',
+  },
   module: {
     rules: [
       {
